完成 production 寄信設定
