+++ app/controllers/admin_controller.rb
@@ -1,0 +1,8 @@
+class AdminController < ApplicationController
+  before_action :authenticate_user!
+  before_action :admin_required
+
+  # before_action :xxxxx if is_admin
+  layout "admin"
+
+end
