+++ app/service/order_placing_service.rb
@@ -8,6 +8,6 @@ class OrderPlacingService
     @order.build_item_cache_from_cart(@cart)
     @order.calculate_total!(@cart)
     @cart.clean!
-    # OrderMailer.notify_order_placed(@order).deliver
+    OrderMailer.notify_order_placed(@order).deliver
   end
 end
