+++ config/environments/production.rb
@@ -82,4 +82,14 @@ Rails.application.configure do
     # Pay2go.integration_mode = :production
     Pay2go.integration_mode = :development
   end
+
+  config.action_mailer.delivery_method = :smtp
+  config.action_mailer.smtp_settings = {
+      port: 587,
+      address: "smtp.mailgun.com",
+      user_name: ENV["mailgun_user"],
+      password: ENV["mailgun_secret"],
+      domain: "sandboxb91cb02592444783ba82e83abd61add2.mailgun.org",
+      authentication: :plain,
+  }
 end
